Enhanced OpenAI Compatible model
I want to be able to use other types of models that OpenAI offers without
modifying the `openai.py` client -- which hardcodes its models.

This diff makes this possible by the following changes:

- Updated `is_configured_properly()` method to check for the OpenAI API key in
environment variables when not found in the config.
- Added a check to ensure that we do not leak the OPENAI_API_KEY to the upstream
openai compatible provider, while checking that we're using the OpenAPI API key
with a compatible model only when accompanying it with an explicit base url to
OpenAI's API url.
- Modified `get_config()` to let us retrieve values from a list of strings while
maintaining backward compatibility with previous releases.